+++ wp-testing.php
@@ -3,7 +3,7 @@
  * Plugin Name: Wp-testing
  * Plugin URI: http://wordpress.org/extend/plugins/wp-testing/
  * Description: Helps to create psychological tests.
- * Version: 0.9.5
+ * Version: 0.10
  * Author: Alexander Ustimenko
  * Author URI: http://ustimen.co
  * License: GPL3
@@ -14,4 +14,4 @@
 require_once dirname(__FILE__) . '/src/WordPressFacade.php';
 require_once dirname(__FILE__) . '/src/Facade.php';
 
-new WpTesting_Facade(new WpTesting_WordPressFacade(__FILE__));
+$WpTesting_Facade = new WpTesting_Facade(new WpTesting_WordPressFacade(__FILE__));